% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_value.R
\name{summarize_value}
\alias{summarize_value}
\title{Summarizes the values of the mean of any variable of the data set, and groups it by any variable.}
\usage{
summarize_value(data, vars, s_var)
}
\arguments{
\item{data}{Must be in a data set that contains numerical data that can be mutated.}

\item{vars}{The variable that is being summarized via finding the mean. Must be numerical.}

\item{s_var}{The variable that the summarized data is being grouped by. Must be categorical.}
}
\value{
mutation Returns the data set with just the selected categorical variable and the means of the selected numeric variable.
}
\description{
Uses the inputted data and variable to group by the chosen variable to give a summation of the mean for each categorical variable from the numerical variable inputted.
}
